Factor out citow/itow from ToddCoxeter
